+++ persistent/Database/Persist/EntityDef.hs
@@ -13,6 +13,7 @@ module Database.Persist.EntityDef
     , getEntityFieldsDatabase
     , getEntityForeignDefs
     , getEntityUniques
+    , getEntityUniquesNoPrimaryKey
     , getEntityId
     , getEntityIdField
     , getEntityKeyFields
@@ -39,22 +40,39 @@ import Database.Persist.EntityDef.Internal
 import Database.Persist.FieldDef
 
 import Database.Persist.Types.Base
-    ( UniqueDef
+    ( UniqueDef(..)
     , ForeignDef
     , entityKeyFields
     )
 import Database.Persist.Names
 
--- | Retrieve the list of 'UniqueDef' from an 'EntityDef'. This currently does
--- not include a @Primary@ key, if one is defined. A future version of
--- @persistent@ will include a @Primary@ key among the 'Unique' constructors for
--- the 'Entity'.
+-- | Retrieve the list of 'UniqueDef' from an 'EntityDef'. This does not include
+-- a @Primary@ key, if one is defined. A future version of @persistent@ will
+-- include a @Primary@ key among the 'Unique' constructors for the 'Entity'.
+--
+-- @since 2.14.0.0
+getEntityUniquesNoPrimaryKey
+    :: EntityDef
+    -> [UniqueDef]
+getEntityUniquesNoPrimaryKey ed =
+    filter isNotPrimaryKey $ entityUniques ed
+  where
+    isNotPrimaryKey ud =
+        let
+            constraintName = unConstraintNameHS $ uniqueHaskell ud
+         in
+            constraintName /= unEntityNameHS (getEntityHaskellName ed) <> "PrimaryKey"
+
+-- | Retrieve the list of 'UniqueDef' from an 'EntityDef'.  As of version 2.14,
+-- this will also include the primary key on the entity, if one is defined. If
+-- you do not want the primary key, see 'getEntityUniquesNoPrimaryKey'.
 --
 -- @since 2.13.0.0
 getEntityUniques
     :: EntityDef
     -> [UniqueDef]
-getEntityUniques = entityUniques
+getEntityUniques =
+    entityUniques
 
 -- | Retrieve the Haskell name of the given entity.
 --