+++ persistent/Database/Persist/Quasi/Internal.hs
@@ -51,6 +51,7 @@ module Database.Persist.Quasi.Internal
 
 import Prelude hiding (lines)
 
+import Control.Monad
 import Control.Applicative (Alternative((<|>)))
 import Data.Char (isLower, isSpace, isUpper, toLower)
 import Data.List (find, foldl')
@@ -461,7 +462,7 @@ unbindCompositeDef :: CompositeDef -> UnboundCompositeDef
 unbindCompositeDef cd =
     UnboundCompositeDef
         { unboundCompositeCols =
-            NEL.toList $ fmap fieldHaskell (compositeFields cd)
+            fmap fieldHaskell (compositeFields cd)
         , unboundCompositeAttrs =
             compositeAttrs cd
         }
@@ -984,10 +985,16 @@ takeConstraint ps entityName defs (n :| rest) =
                     Just $ pure (takeForeign ps entityName rest)
                 }
         "Primary" ->
-            mempty
-                { entityConstraintDefsPrimaryComposite =
-                    SetOnce (takeComposite (unboundFieldNameHS <$> defs) rest)
-                }
+            let
+                unboundComposite =
+                    takeComposite (unboundFieldNameHS <$> defs) rest
+            in
+                mempty
+                    { entityConstraintDefsPrimaryComposite =
+                        SetOnce unboundComposite
+                    , entityConstraintDefsUniques =
+                        Just $ pure $ compositeToUniqueDef entityName defs unboundComposite
+                    }
         "Id" ->
             mempty
                 { entityConstraintDefsIdField =
@@ -1067,7 +1074,7 @@ takeId ps entityName texts =
 --
 -- @since.2.13.0.0
 data UnboundCompositeDef = UnboundCompositeDef
-    { unboundCompositeCols :: [FieldNameHS]
+    { unboundCompositeCols :: NonEmpty FieldNameHS
     -- ^ The field names for the primary key.
     --
     -- @since 2.13.0.0
@@ -1079,6 +1086,31 @@ data UnboundCompositeDef = UnboundCompositeDef
     }
     deriving (Eq, Ord, Show, Lift)
 
+compositeToUniqueDef :: EntityNameHS -> [UnboundFieldDef] -> UnboundCompositeDef -> UniqueDef
+compositeToUniqueDef entityName fields UnboundCompositeDef {..} =
+    UniqueDef
+        { uniqueHaskell =
+            ConstraintNameHS (unEntityNameHS entityName <> "PrimaryKey")
+        , uniqueDBName =
+            ConstraintNameDB "primary_key"
+        , uniqueFields =
+            fmap (\hsName -> (hsName, getDbNameFor hsName)) unboundCompositeCols
+        , uniqueAttrs =
+            unboundCompositeAttrs
+        }
+  where
+    getDbNameFor hsName =
+        case mapMaybe (matchHsName hsName) fields of
+            [] ->
+                error "Unable to find `hsName` in fields"
+            (a : _) ->
+                a
+    matchHsName hsName UnboundFieldDef {..} = do
+        guard $ unboundFieldNameHS == hsName
+        pure unboundFieldNameDB
+
+
+
 takeComposite
     :: [FieldNameHS]
     -> [Text]
@@ -1086,11 +1118,17 @@ takeComposite
 takeComposite fields pkcols =
     UnboundCompositeDef
         { unboundCompositeCols =
-            map (getDef fields) cols
+            fmap (getDef fields) neCols
         , unboundCompositeAttrs =
             attrs
         }
   where
+    neCols =
+        case NEL.nonEmpty cols of
+            Nothing ->
+                error "No fields provided for primary key"
+            Just xs ->
+                xs
     (cols, attrs) = break ("!" `T.isPrefixOf`) pkcols
     getDef [] t = error $ "Unknown column in primary key constraint: " ++ show t
     getDef (d:ds) t